Big modification of the global structure.
The goal was too delete some useless dependences to Triton.h
and pin.H.

The other key idea is too lighten Triton.h, it should contain
only the necessary include and definition/declaration useful in
core.cpp.

How do we achieve these goals ?

For pin.H, it's pretty simple most of the types define by pin.H (UINT64,
ADDRINT, ...) are here only for handle Microsoft and Unix platforms.
Right now, we focus only on Unix so we just need to use define in
stdint.h (uint64_h).

For Triton.h, it's trickier but the solution is ,the most often, to
define some headers or tweak the existing one. And then include this
"new" header in Triton.h.