Fix issues 13: change containers in SymbolicEngine.
Each list has different purposes and complexity requirements.
So for most of them other kind of containers are most suited.

memoryReference becomes a map associating an address to a reference ID.
symVarMemoryReference becomes also a map associating an address to
a symVarMemory ID.
This change is interesting because we will seek an element in these
maps more often than adding a new element.

Complexity are:
    |List | Map
---------------------
add |O(1) | O(nlog(n))
seek|O(n) | O(log(n))

symbolicList become a symbolicVector, it allows the method getElementFromId
to be in constant time.