End Fix issue #13
To fit our needs, a collision in memoryReference must result to an
overwrite of the ancient value. The insert method of std::map
doesn't do this, so we use the overloading of the [] operator
instead.

symVarMemoryReference becomes a map from a symVar ID to a memory
address. This way correspond to its future use.

I change the signature to quickly of isMemoryReference in the last
commit, it was an error.
The value -1 is used as a NIL value and so this method can
definitively not return an unsigned int!