Fix extra quote raising warning on windows:
CMake_ Warning (dev) at src/libtriton/CMakeLists.txt:21 (if):
Policy CMP0054 is not set: Only interpret if() arguments as variables or
keywords when unquoted. Run "cmake --help-policy CMP0054" for policy
details. Use the cmake_policy command to set the policy and suppress
this warning.

Quoted variables like "MSVC" will no longer be dereferenced when the
policy is set to NEW. Since the policy is not set the OLD behavior will
be used.