Removing the SYSCALL namespace
Triton tried to generate a SYSCALL enum. For that, during the compile Triton
needs to read the syscalls id table of the Kernel and then it generates its enum
according to the Kernel spec during the compile. However, it means that the
Triton's enum is specific to the Kernel version where it was compiled. First, as
we know, syscalls id table change from a kernel version to another and it's hard
to keep a correct enum for every version. Second, as Triton is now a cross
architecture emulator, when analyzing an aarch64 binary, the SYSCALL enum is
not the same as the one generated during the compile for x64. Third, Triton does
not need this enum to work. The enum is just provided as information.

For all these reasons, we removed this enums.